requirements.txt
Flask==2.1.0
Werkzeug==2.1.0
transformers==4.6.0
torch>=1.11.0
tokenizers==0.10.3
gunicorn==20.1.0
six==1.16.0
python-dotenv==1.0.1
tweepy==4.14.0